good_practices/pyproject.toml
[project]
name = "good_practices"
version = "0.0.1"
authors = [{ name = "agustin.panizza@sunnydata.ai" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
    "org-logging",
    "pyspark>=4.1.1",
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
    "zstandard>=0.25.0",
    "ruff>=0.14.13",
]

[project.scripts]
main = "good_practices.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
org-logging = { path = "org_logging-0.1.0+1768832188192251944-py3-none-any.whl" }

[tool.black]
line-length = 125
